Adding section Hero
