+++ src/css/mobile-menu.css
@@ -34,14 +34,14 @@
         min-height: 100%;
         display: flex;
         flex-direction: column;
-        background-color: #FFFFFF;
+        background-color: #f0f0f0;
         background-image: url('../img/menu/decor-bg-mob.png');
         background-size: cover;
     }
 
     .sidebar-header-container {
         display: flex;
-        padding: 36px 20px;
+        padding: 16px 20px;
         justify-content: space-between;
         max-width: 375px;
         margin: 0 auto;
@@ -76,7 +76,8 @@
         flex-direction: column;
         align-items: center;
         row-gap: 24px;
-        padding: 30vh 101px 50px 101px;
+        justify-content: center;
+        flex: 1 1 auto;
     }
 
     .sidebar-nav-link {
