+++ src/js/hero.js
@@ -1,0 +1,15 @@
+ // Smooth scroll to sections
+document.querySelectorAll('.smooth-scroll').forEach(anchor => {
+  anchor.addEventListener('click', function (e) {
+    e.preventDefault();
+
+    const targetId = this.getAttribute('href');
+    const targetElement = document.querySelector(targetId);
+
+    if (targetElement) {
+      targetElement.scrollIntoView({
+        behavior: 'smooth', // Smooth scroll
+      });
+    }
+  });
+});
