+++ src/partials/hero.html
@@ -32,7 +32,6 @@
       </div>
     </div>
     <div class="hero-fstack-container">
-      <!-- <div class="hero-social"> -->
       <div class="hero-social-container">
         <ul class="hero-social-ul">
           <li class="hero-social-li">
@@ -79,18 +78,7 @@
           </span>
           <span class="hero-text">ORDER THE PROJECT</span>
         </a>
-        <!-- </div> -->
       </div>
-      <!-- <div class="hero-fstack-order">
-        <a href="#work-together" class="hero-custom-button smooth-scroll">
-          <span class="hero-arrow">
-            <svg class="hero-icon">
-              <use href="../img/icons/sprite.svg#icon-right-hero"></use>
-            </svg>
-          </span>
-          <span class="hero-text">ORDER THE PROJECT</span>
-        </a>
-      </div> -->
     </div>
     <div class="hero-moto">
       I WORK WITH PEOPLE ALL OVER THE WORLD TO CREATE
